updates conversation routes to use query parameters for ID
