--- src/app/api/chat/conversations/[id]/route.ts
@@ -1,106 +1,0 @@
-import { NextRequest, NextResponse } from "next/server";
-
-export async function GET(
-  req: NextRequest,
-  { params }: { params: { id: string } }
-) {
-  try {
-    // Get token from Authorization header
-    const authHeader = req.headers.get("Authorization");
-    if (!authHeader) {
-      return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
-    }
-    
-    const token = authHeader.replace("Bearer ", "");
-    
-    const { id } = await params;
-    
-    const response = await fetch(`${process.env.API_URL}/chat/conversations/${id}`, {
-      method: "GET",
-      headers: {
-        "Content-Type": "application/json",
-        "Authorization": `Bearer ${token}`
-      },
-    });
-
-    const data = await response.json();
-    return NextResponse.json(data);
-  } catch (error) {
-    console.error("Error fetching conversation:", error);
-    return NextResponse.json(
-      { error: "Failed to fetch conversation" },
-      { status: 500 }
-    );
-  }
-}
-
-export async function PUT(
-  req: NextRequest,
-  { params }: { params: { id: string } }
-) {
-  try {
-    // Get token from Authorization header
-    const authHeader = req.headers.get("Authorization");
-    if (!authHeader) {
-      return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
-    }
-    
-    const token = authHeader.replace("Bearer ", "");
-    
-    const { id } = await params;
-    
-    const body = await req.json();
-    
-    const response = await fetch(`${process.env.API_URL}/chat/conversations/${id}`, {
-      method: "PUT",
-      headers: {
-        "Content-Type": "application/json",
-        "Authorization": `Bearer ${token}`
-      },
-      body: JSON.stringify(body)
-    });
-
-    const data = await response.json();
-    return NextResponse.json(data);
-  } catch (error) {
-    console.error("Error updating conversation:", error);
-    return NextResponse.json(
-      { error: "Failed to update conversation" },
-      { status: 500 }
-    );
-  }
-}
-
-export async function DELETE(
-  req: NextRequest,
-  { params }: { params: { id: string } }
-) {
-  try {
-    // Get token from Authorization header
-    const authHeader = req.headers.get("Authorization");
-    if (!authHeader) {
-      return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
-    }
-    
-    const token = authHeader.replace("Bearer ", "");
-    
-    const { id } = await params;
-    
-    const response = await fetch(`${process.env.API_URL}/chat/conversations/${id}`, {
-      method: "DELETE",
-      headers: {
-        "Content-Type": "application/json",
-        "Authorization": `Bearer ${token}`
-      },
-    });
-
-    const data = await response.json();
-    return NextResponse.json(data);
-  } catch (error) {
-    console.error("Error deleting conversation:", error);
-    return NextResponse.json(
-      { error: "Failed to delete conversation" },
-      { status: 500 }
-    );
-  }
-}
