+++ src/app/api/chat/conversations/route.ts
@@ -10,7 +10,26 @@ export async function GET(req: NextRequest) {
     
     const token = authHeader.replace("Bearer ", "");
     
-    const response = await fetch(`${process.env.API_URL}/chat/conversations`, {
+    // Get conversation ID from query parameters
+    const { searchParams } = new URL(req.url);
+    const id = searchParams.get("id");
+    
+    // If no ID is provided, fetch all conversations
+    if (!id) {
+      const response = await fetch(`${process.env.API_URL}/chat/conversations`, {
+        method: "GET",
+        headers: {
+          "Content-Type": "application/json",
+          "Authorization": `Bearer ${token}`
+        },
+      });
+  
+      const data = await response.json();
+      return NextResponse.json(data);
+    }
+    
+    // If ID is provided, fetch specific conversation
+    const response = await fetch(`${process.env.API_URL}/chat/conversations/${id}`, {
       method: "GET",
       headers: {
         "Content-Type": "application/json",
@@ -21,9 +40,9 @@ export async function GET(req: NextRequest) {
     const data = await response.json();
     return NextResponse.json(data);
   } catch (error) {
-    console.error("Error fetching conversations:", error);
+    console.error("Error fetching conversation(s):", error);
     return NextResponse.json(
-      { error: "Failed to fetch conversations" },
+      { error: "Failed to fetch conversation(s)" },
       { status: 500 }
     );
   }
@@ -55,6 +74,89 @@ export async function POST(req: NextRequest) {
     console.error("Error creating conversation:", error);
     return NextResponse.json(
       { error: "Failed to create conversation" },
+      { status: 500 }
+    );
+  }
+}
+
+export async function PUT(req: NextRequest) {
+  try {
+    // Get token from Authorization header
+    const authHeader = req.headers.get("Authorization");
+    if (!authHeader) {
+      return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
+    }
+    
+    const token = authHeader.replace("Bearer ", "");
+    
+    // Get conversation ID from query parameters
+    const { searchParams } = new URL(req.url);
+    const id = searchParams.get("id");
+    
+    if (!id) {
+      return NextResponse.json(
+        { error: "Conversation ID is required" },
+        { status: 400 }
+      );
+    }
+    
+    const body = await req.json();
+    
+    const response = await fetch(`${process.env.API_URL}/chat/conversations/${id}`, {
+      method: "PUT",
+      headers: {
+        "Content-Type": "application/json",
+        "Authorization": `Bearer ${token}`
+      },
+      body: JSON.stringify(body)
+    });
+
+    const data = await response.json();
+    return NextResponse.json(data);
+  } catch (error) {
+    console.error("Error updating conversation:", error);
+    return NextResponse.json(
+      { error: "Failed to update conversation" },
+      { status: 500 }
+    );
+  }
+}
+
+export async function DELETE(req: NextRequest) {
+  try {
+    // Get token from Authorization header
+    const authHeader = req.headers.get("Authorization");
+    if (!authHeader) {
+      return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
+    }
+    
+    const token = authHeader.replace("Bearer ", "");
+    
+    // Get conversation ID from query parameters
+    const { searchParams } = new URL(req.url);
+    const id = searchParams.get("id");
+    
+    if (!id) {
+      return NextResponse.json(
+        { error: "Conversation ID is required" },
+        { status: 400 }
+      );
+    }
+    
+    const response = await fetch(`${process.env.API_URL}/chat/conversations/${id}`, {
+      method: "DELETE",
+      headers: {
+        "Content-Type": "application/json",
+        "Authorization": `Bearer ${token}`
+      },
+    });
+
+    const data = await response.json();
+    return NextResponse.json(data);
+  } catch (error) {
+    console.error("Error deleting conversation:", error);
+    return NextResponse.json(
+      { error: "Failed to delete conversation" },
       { status: 500 }
     );
   }
